[ALPS02407758] DVFS: hopping workaround cast ul
[Detail] hopping workaround cast ul

[Solution] hopping workaround cast ul

[Feature] DVFS (Dynamic Voltage and Frequency Scaling)

MTK-Commit-Id: 04484341a13613a45bec806abf8dffda507a100c

CR-Id: ALPS02407758

+++ drivers/misc/mediatek/base/power/mt6797/mt_cpufreq.c
@@ -642,7 +642,7 @@ cpufreq_write(addr, (cpufreq_read(addr) & ~(_BITMASK_(mask))) | _BITS_(mask, val
 #define cpufreq_read_armpll(addr)				mt6797_0x1001AXXX_reg_read((unsigned long)addr)
 #define cpufreq_write_armpll(addr, val)			mt6797_0x1001AXXX_reg_write((unsigned long)addr, val)
 #define cpufreq_write_mask_armpll(addr, mask, val) \
-mt6797_0x1001AXXX_reg_set(addr, (_BITMASK_(mask)), val)
+mt6797_0x1001AXXX_reg_set((unsigned long)addr, (_BITMASK_(mask)), val)
 
 /*
  * LOCK
